pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = ["Belg PC <tomsomoza911@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.5.0"
requests = "^2.32.3"
polars = "^1.23.0"
python-bitvavo-api = "^1.4.2"
gradio = "^5.20.1"
matplotlib = "^3.10.1"
pyperclip = "^1.9.0"
loguru = "^0.7.3"
rich = "^13.9.4"
pytest = "^8.3.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
